+++ src/elementos/ElementoEstatico.java
@@ -1,8 +1,8 @@
-package cenario;
-
-import jogoCataFrutas.Elemento;
+package elementos;
 
 public abstract class ElementoEstatico extends Elemento {
+	public ElementoEstatico() {  }
+	
     public ElementoEstatico(String nome, int posX, int posY) {
         super(nome, posX, posY);
     }
