+++ src/elementos/Pedra.java
@@ -1,6 +1,8 @@
-package cenario;
+package elementos;
 
 public class Pedra extends ElementoEstatico {
+	public Pedra() {  }
+	
     public Pedra(String nome, int posX, int posY) {
         super(nome, posX, posY);
     }
